requirements.txt
prettytable
mysql-connector-python
python-dotenv
